requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
python-multipart==0.0.12
pymupdf==1.24.10
python-docx==1.1.2
jinja2==3.1.4
aiofiles==23.2.1
python-dotenv==1.0.1
authlib==1.3.1
pymongo==4.8.0
itsdangerous==2.2.0